add stream conversations to example app
