+++ example/src/main/java/org/xmtp/android/example/conversation/ConversationDetailViewModel.kt
@@ -4,22 +4,19 @@ import androidx.annotation.UiThread
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingCommand
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.emptyFlow
-import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.launch
 import org.xmtp.android.example.ClientManager
+import org.xmtp.android.example.extension.flowWhileShared
+import org.xmtp.android.example.extension.stateFlow
 import org.xmtp.android.library.Conversation
 import org.xmtp.android.library.DecodedMessage
 
@@ -88,65 +85,38 @@ class ConversationDetailViewModel(private val savedStateHandle: SavedStateHandle
             emptyFlow()
         }
 
-@UiThread
-fun sendMessage(body: String): StateFlow<SendMessageState> {
-    val flow = MutableStateFlow<SendMessageState>(SendMessageState.Loading)
-    viewModelScope.launch(Dispatchers.IO) {
-        try {
-            conversation?.send(body)
-            flow.value = SendMessageState.Success
-        } catch (e: Exception) {
-            flow.value = SendMessageState.Error(e.localizedMessage.orEmpty())
-        }
-    }
-    return flow
-}
-
-sealed class UiState {
-    data class Loading(val listItems: List<MessageListItem>?) : UiState()
-    data class Success(val listItems: List<MessageListItem>) : UiState()
-    data class Error(val message: String) : UiState()
-}
-
-sealed class SendMessageState {
-    object Loading : SendMessageState()
-    object Success : SendMessageState()
-    data class Error(val message: String) : SendMessageState()
-}
-
-sealed class MessageListItem(open val id: String, val itemType: Int) {
-    companion object {
-        const val ITEM_TYPE_MESSAGE = 1
-    }
-
-    data class Message(override val id: String, val message: DecodedMessage) :
-        MessageListItem(id, ITEM_TYPE_MESSAGE)
-}
-
-@OptIn(ExperimentalCoroutinesApi::class)
-private fun <T> Flow<T>.flowWhileShared(
-    subscriptionCount: StateFlow<Int>,
-    started: SharingStarted,
-): Flow<T> {
-    return started.command(subscriptionCount)
-        .distinctUntilChanged()
-        .flatMapLatest {
-            when (it) {
-                SharingCommand.START -> this
-                SharingCommand.STOP, SharingCommand.STOP_AND_RESET_REPLAY_CACHE -> emptyFlow()
+    @UiThread
+    fun sendMessage(body: String): StateFlow<SendMessageState> {
+        val flow = MutableStateFlow<SendMessageState>(SendMessageState.Loading)
+        viewModelScope.launch(Dispatchers.IO) {
+            try {
+                conversation?.send(body)
+                flow.value = SendMessageState.Success
+            } catch (e: Exception) {
+                flow.value = SendMessageState.Error(e.localizedMessage.orEmpty())
             }
         }
-}
-
-private fun <T> stateFlow(
-    scope: CoroutineScope,
-    initialValue: T,
-    producer: (subscriptionCount: StateFlow<Int>) -> Flow<T>,
-): StateFlow<T> {
-    val state = MutableStateFlow(initialValue)
-    scope.launch {
-        producer(state.subscriptionCount).collect(state)
+        return flow
     }
-    return state.asStateFlow()
-}
+
+    sealed class UiState {
+        data class Loading(val listItems: List<MessageListItem>?) : UiState()
+        data class Success(val listItems: List<MessageListItem>) : UiState()
+        data class Error(val message: String) : UiState()
+    }
+
+    sealed class SendMessageState {
+        object Loading : SendMessageState()
+        object Success : SendMessageState()
+        data class Error(val message: String) : SendMessageState()
+    }
+
+    sealed class MessageListItem(open val id: String, val itemType: Int) {
+        companion object {
+            const val ITEM_TYPE_MESSAGE = 1
+        }
+
+        data class Message(override val id: String, val message: DecodedMessage) :
+            MessageListItem(id, ITEM_TYPE_MESSAGE)
+    }
 }
